+++ .editorconfig
@@ -1,0 +1,12 @@
+root = true
+
+[src/**.{js,css,scss}]
+end_of_line = lf
+insert_final_newline = true
+indent_style = space
+indent_size = 4
+
+
+[{package.json, webpack.config.js}]
+indent_style = space
+indent_size = 4
