+++ .gitignore
@@ -2,7 +2,7 @@
 .DS_Store
 Thumbs.db
 
-# IDEs 
+# IDEs
 .buildpath
 .project
 .settings/
@@ -12,7 +12,8 @@ Thumbs.db
 nbproject/
 _notes/
 
-# composer related
+# composer/node related
+node_modules/
 vendor/
 
 # build
