+++ package.json
@@ -1,0 +1,35 @@
+{
+    "name": "contao-multicolumnwizard-bundle",
+    "main": "src/Resources/public/js/multicolumnwizard_be.js",
+    "scripts": {
+        "build": "webpack --mode=production",
+        "start": "webpack --watch --mode=development"
+    },
+    "license": "LGPL-3.0-or-later",
+    "bugs": {
+        "url": "https://github.com/menatwork/MultiColumnWizard/issues"
+    },
+    "homepage": "https://github.com/menatwork/MultiColumnWizard#readme",
+    "dependencies": {},
+    "devDependencies": {
+        "@babel/core": "^7.7.2",
+        "@babel/preset-env": "^7.7.1",
+        "babel-loader": "^8.0.6",
+        "css-loader": "^3.2.0",
+        "git-revision-webpack-plugin": "^3.0.4",
+        "husky": "^3.0.9",
+        "mini-css-extract-plugin": "^0.8.0",
+        "node-sass": "^4.13.0",
+        "postcss-loader": "^3.0.0",
+        "postcss-preset-env": "^6.7.0",
+        "sass-loader": "^8.0.0",
+        "webpack": "^4.41.2",
+        "webpack-cli": "^3.3.10"
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "npm run build"
+        }
+    },
+    "browserslist": "> 0.25%, not dead"
+}
