+++ src/Resources/assets/css/multicolumnwizard.scss
@@ -1,0 +1,63 @@
+.multicolumnwizard {
+    margin-top: 5px;
+}
+
+.multicolumnwizard h3 {
+    padding: 0;
+    overflow: hidden;
+}
+
+.multicolumnwizard .tl_textarea {
+    width: auto;
+    height: auto;
+}
+
+.multicolumnwizard .tl_select,
+.multicolumnwizard .tl_mselect {
+    width: auto;
+}
+
+.multicolumnwizard .tl_checkbox_single_container {
+    margin-top: 0;
+}
+
+.multicolumnwizard .hidelabel h3,
+.multicolumnwizard .hidelabel label {
+    display: none;
+}
+
+.multicolumnwizard label {
+    font-weight: normal;
+    float: left;
+}
+
+.multicolumnwizard input.tl_checkbox {
+    float: left;
+}
+
+.multicolumnwizard td.col_last {
+    padding-right: 0;
+}
+
+.multicolumnwizard a[rel="move"] {
+    cursor: move;
+}
+
+.multicolumnwizard .w50 {
+    float: none;
+    width: auto;
+    height: auto;
+}
+
+.multicolumnwizard .tl_checkbox_container label {
+    margin: 0;
+    padding-left: 5px;
+}
+
+.multicolumnwizard a.rotate img {
+    animation: mcw-spin .4s linear infinite;
+}
+
+@keyframes mcw-spin {
+    100% {transform:rotate(360deg);}
+}
