+++ src/Resources/public/css/multicolumnwizard_src.css
@@ -1,81 +1,12 @@
-/**
- * Contao Open Source CMS
- *
+/*!
+ * @name        contao-multicolumnwizard-bundle
+ * @version     3.4.2-1-gd4a0b33
+ * @copyright   Andreas Schempp 2011
+ * @copyright   certo web & design GmbH 2011
  * @copyright   MEN AT WORK 2013
  * @package     MultiColumnWizard
- * @license     GNU/LGPL
- * @info        tab is set to 4 whitespaces
+ * @license     LGPL-3.0-or-later
+ * @file        main
  */
+.multicolumnwizard{margin-top:5px}.multicolumnwizard h3{padding:0;overflow:hidden}.multicolumnwizard .tl_textarea{width:auto;height:auto}.multicolumnwizard .tl_select,.multicolumnwizard .tl_mselect{width:auto}.multicolumnwizard .tl_checkbox_single_container{margin-top:0}.multicolumnwizard .hidelabel h3,.multicolumnwizard .hidelabel label{display:none}.multicolumnwizard label{font-weight:normal;float:left}.multicolumnwizard input.tl_checkbox{float:left}.multicolumnwizard td.col_last{padding-right:0}.multicolumnwizard a[rel="move"]{cursor:move}.multicolumnwizard .w50{float:none;width:auto;height:auto}.multicolumnwizard .tl_checkbox_container label{margin:0;padding-left:5px}.multicolumnwizard a.rotate img{-webkit-animation:mcw-spin .4s linear infinite;animation:mcw-spin .4s linear infinite}@-webkit-keyframes mcw-spin{100%{-webkit-transform:rotate(360deg);transform:rotate(360deg)}}@keyframes mcw-spin{100%{-webkit-transform:rotate(360deg);transform:rotate(360deg)}}
 
-.multicolumnwizard {
-    margin-top: 5px;
-}
-
-.multicolumnwizard h3 {
-    padding: 0;
-    overflow: hidden;
-}
-
-.multicolumnwizard .tl_textarea {
-    width: auto;
-    height: auto;
-}
-
-.multicolumnwizard .tl_select, 
-.multicolumnwizard .tl_mselect {
-    width: auto;
-}
-
-.multicolumnwizard .tl_checkbox_single_container {
-    margin-top: 0;
-}
-
-.multicolumnwizard .hidelabel h3,
-.multicolumnwizard .hidelabel label {
-    display: none;
-}
-
-.multicolumnwizard label {
-    font-weight: normal;
-    float: left;
-}
-
-.multicolumnwizard input.tl_checkbox {
-    float: left;
-}
-
-.multicolumnwizard td.col_last {
-    padding-right: 0;
-}
-
-.multicolumnwizard a[rel="move"] {
-    cursor: move;
-}
-
-
-.multicolumnwizard .w50 {
-    float: none;
-    width: auto;
-    height: auto;
-}
-
-.multicolumnwizard .tl_checkbox_container label {
-    margin: 0;
-    padding-left: 5px;
-}
-
-.multicolumnwizard a.rotate img {
-    -webkit-animation: mcw-spin .4s linear infinite;
-    -moz-animation: mcw-spin .4s linear infinite;
-    animation: mcw-spin .4s linear infinite;
-}
-
-@-moz-keyframes mcw-spin {
-    100% {-moz-transform: rotate(360deg);}
-}
-@-webkit-keyframes mcw-spin {
-    100% {-webkit-transform: rotate(360deg);}
-}
-@keyframes mcw-spin {
-    100% {-webkit-transform: rotate(360deg); transform:rotate(360deg);}
-}
